Move bootstrap input components to inputs folder
