+++ src/components/bootstrap/index.js
@@ -1,1 +1,1 @@
-export InputGroup from './InputGroup'
+export * from './inputs'
