+++ src/components/bootstrap/inputs/index.js
@@ -1,0 +1,1 @@
+export InputGroup from './InputGroup'
