+++ stories/components/bootstrap/inputs/InputGroup.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { compose, withState } from 'recompose'
 import { storiesOf } from '@kadira/storybook'
 
-import { InputGroup, Text } from '../../../src'
+import { Icon, InputGroup, Text } from '../../../../src'
 
 const TextWithState = compose(
   withState('value', 'setValue', props => props.value),
@@ -26,24 +26,31 @@ export default () => {
         />
       </InputGroup>
     ))
-    .add('Left string addon', () => (
-      <InputGroup leftAddon="@">
+    .add('Left addon', () => (
+      <InputGroup
+        leftAddon={<Icon icon="code" />}
+      >
         <TextWithState
           className="form-control"
           value="Testing"
         />
       </InputGroup>
     ))
-    .add('Right string addon', () => (
-      <InputGroup rightAddon="@">
+    .add('Right addon', () => (
+      <InputGroup
+        rightAddon={<Icon icon="code" />}
+      >
         <TextWithState
           className="form-control"
           value="Testing"
         />
       </InputGroup>
     ))
-    .add('Left & Right string addons', () => (
-      <InputGroup leftAddon="@" rightAddon="@">
+    .add('Left & Right addons', () => (
+      <InputGroup
+        leftAddon={<Icon icon="code" />}
+        rightAddon={<Icon icon="times" />}
+      >
         <TextWithState
           className="form-control"
           value="Testing"
@@ -51,7 +58,12 @@ export default () => {
       </InputGroup>
     ))
     .add('Left multi addon', () => (
-      <InputGroup leftAddon={['@', 'www.example.com']}>
+      <InputGroup
+        leftAddon={[
+          <Icon icon="user" />,
+          <Icon icon="code" />,
+        ]}
+      >
         <TextWithState
           className="form-control"
           value="Testing"
@@ -59,7 +71,12 @@ export default () => {
       </InputGroup>
     ))
     .add('Right multi addon', () => (
-      <InputGroup rightAddon={['@', 'www.example.com']}>
+      <InputGroup
+        rightAddon={[
+          <Icon icon="user" />,
+          <Icon icon="code" />,
+        ]}
+      >
         <TextWithState
           className="form-control"
           value="Testing"
