+++ stories/components/bootstrap/inputs/index.js
@@ -1,0 +1,5 @@
+import InputGroup from './InputGroup'
+
+export default () => {
+  InputGroup()
+}
